Revert "All bad"
This reverts commit ca0e3ea751e19958d592133e0a4ec8cb3a3c9be3.

+++ Flower_shop/Controllers/AdminPlatformController.cs
@@ -62,6 +62,7 @@ namespace Flower_shop.Controllers
                 TypesName = typeName
             };
 
+
             return View(productView);
         }
         [HttpPost]
@@ -124,7 +125,7 @@ namespace Flower_shop.Controllers
             return View();
         }
         [HttpPost]
-        public async Task<IActionResult> ProductDelete(long id)
+        public async Task<IActionResult> ProductDelete(int id)
         {
             await _productRepository.RemoveByIdAsync(id);
 
@@ -136,7 +137,7 @@ namespace Flower_shop.Controllers
             return View();
         }
         [HttpPost]
-        public async Task<IActionResult> TypeProductDelete(long typeId)
+        public async Task<IActionResult> TypeProductDelete(int typeId)
         {
             await _typeProductRepository.RemoveTypeProductAsync(typeId);
